requirements-dev.txt
django-debug-toolbar==1.8
django-extensions==1.9.6